requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
asyncio==3.4.3
attrs==21.4.0
chardet==4.0.0
idna==3.3
multidict==6.0.2
numpy==1.23.0
pandas==1.4.3
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
schedule==1.1.0
six==1.16.0
table2ascii==0.3.0
typing_extensions==4.2.0
yarl==1.7.2
